cursor-api-proxy/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.18.1
python-dotenv==1.0.0
pydantic==2.5.0
